--- a/src/core/CMakeLists.txt
+++ b/src/core/CMakeLists.txt
@@ -15,7 +15,25 @@ cxx_link(dfly_core base absl::flat_hash_map absl::str_format redis_lib TRDP::lua
 add_executable(dash_bench dash_bench.cc)
 cxx_link(dash_bench dfly_core redis_test_lib)
 
-cxx_test(dfly_core_test dfly_core TRDP::fast_float LABELS DFLY)
+find_library(LIB_PCRE2 NAMES pcre2-8)
+if(LIB_PCRE2)
+  set(PCRE2_LIB ${LIB_PCRE2})
+else()
+  message(STATUS "pcre2-8 not found. Building without PCRE2 support.")
+  set(PCRE2_LIB "")
+endif()
+
+
+find_library(LIB_RE2 NAMES re2)
+if(LIB_RE2)
+  set(RE2_LIB ${LIB_RE2})
+else()
+  message(STATUS "re2 not found. Building without RE2 support.")
+  set(RE2_LIB "")
+endif()
+
+
+cxx_test(dfly_core_test dfly_core TRDP::fast_float ${PCRE2_LIB} ${RE2_LIB} LABELS DFLY)
 cxx_test(compact_object_test dfly_core LABELS DFLY)
 cxx_test(extent_tree_test dfly_core LABELS DFLY)
 cxx_test(dash_test dfly_core file redis_test_lib DATA testdata/ids.txt.zst LABELS DFLY)
@@ -30,3 +48,11 @@ cxx_test(flatbuffers_test dfly_core TRDP::flatbuffers LABELS DFLY)
 cxx_test(bloom_test dfly_core LABELS DFLY)
 cxx_test(allocation_tracker_test dfly_core absl::random_random LABELS DFLY)
 cxx_test(qlist_test dfly_core DATA testdata/list.txt.zst LABELS DFLY)
+
+if(LIB_PCRE2)
+  target_compile_definitions(dfly_core_test PRIVATE USE_PCRE2)
+endif()
+
+if(LIB_RE2)
+  target_compile_definitions(dfly_core_test PRIVATE USE_RE2)
+endif()