--- a/src/core/glob_matcher.h
+++ b/src/core/glob_matcher.h
@@ -3,6 +3,9 @@
 //
 #pragma once
 
+#include <reflex/matcher.h>
+
+#include <string>
 #include <string_view>
 
 namespace dfly {
@@ -16,9 +19,16 @@ class GlobMatcher {
 
   bool Matches(std::string_view str) const;
 
+  // Exposed for testing purposes.
+  static std::string Glob2Regex(std::string_view glob);
+
  private:
-  std::string_view pattern_;
+  mutable reflex::Matcher matcher_;
+
   bool case_sensitive_;
+  bool starts_with_star_ = false;
+  bool ends_with_star_ = false;
+  bool empty_pattern_ = false;
 };
 
 }  // namespace dfly