Also consolidate benchmarking low level routines undeer dfly_core_test

Before this change:
BM_MatchGlob/1000               2164 ns         2164 ns       322600
BM_MatchGlob/10000             21985 ns        21981 ns        31815

After this change:

BM_MatchGlob/1000                137 ns          137 ns      5276350
BM_MatchGlob/10000               437 ns          437 ns      1510854

What's curious is that now matching '*foobar*' on string is faster than searching for 'foobar` using string:find.

<!--
**Commits Must Be Signed and Your PR title must conform to the conventional commit spec**
  * See: https://github.com/dragonflydb/dragonfly/blob/main/CONTRIBUTING.md
  * Please follow the section on `pre-commit hooks`, a linter will validate before you push

  Example PR Title: <type>(<scope>)!: <description>

  * `type` = bug, chore, feat, fix, docs, build, style, refactor, perf, test
  * `!` = OPTIONAL: signals a breaking change
  * `scope` = Optional when `type` is "chore" or "docs"
  * `description` = short description of the change

Examples:

  * chore(examples): Clarify `docker` usage #120
  * docs(readme): Fix Example Links #121
  * feat(ingest)!: Add new ingest #122
  * fix(ingest): Refactor for loop to list comprehension #123
-->